requirements.txt
telethon
python-dotenv
sqlalchemy
aiosqlite
mistralai